+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Breaking Changes
 
 - Dropped Python 3.9 support
+- Removed dependency on pendulum
+  - This means the type `pendulum.DateTime` has been replaced by a `datetime.datetime` everywhere
 
 ### New